+++ app/src/main/java/org/simple/clinic/scanid/ScanSimpleIdUiActions.kt
@@ -1,5 +1,6 @@
 package org.simple.clinic.scanid
 
+import org.simple.clinic.patient.PatientPrefillInfo
 import org.simple.clinic.patient.businessid.Identifier
 import java.util.UUID