+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Use `.editorconfig` for project code style
 - Prefill NHID as `alternateId` in `PatientEntryScreen`
 - Update facility picker search edit text hint
+- Move `PatientPrefillInfo` to `patient` package to resolve coupling
 
 ## Features
 - Add NHID support in `PatientEntryScreen`